Add flow tester
