+++ tools/FlatFlowMigrationCli/Options/MigrateRepoOptions.cs
@@ -3,41 +3,15 @@
 
 using CommandLine;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.DotNet.ProductConstructionService.Client;
 using FlatFlowMigrationCli.Operations;
-using Microsoft.DotNet.DarcLib.VirtualMonoRepo;
 
 namespace FlatFlowMigrationCli.Options;
 
 [Verb("migrate-repo", HelpText = "Onboards a repository to a flat dependency flow")]
-internal class MigrateRepoOptions : Options
+internal class MigrateRepoOptions : FlatFlowMigrationCliOptions
 {
     [Option('r', "repo", Required = true, HelpText = "Name of the repository to migrate (should correspond to a source mapping)")]
     public required string? Mapping { get; init; }
 
-    [Option("pcsUri", Required = false, Default = "https://maestro.dot.net/", HelpText = "PCS base URI, defaults to the Prod PCS")]
-    public required string PcsUri { get; init; }
-
-    [Option("vmr", Required = false, Default = "https://github.com/dotnet/dotnet", HelpText = "URI or path to the VMR. Defaults to https://github.com/dotnet/dotnet")]
-    public required string VmrUri { get; init; }
-
-    [Option("tmp", Required = false, HelpText = "Temporary path where intermediate files are stored (e.g. cloned repos, patch files); defaults to usual TEMP.")]
-    public string? TmpPath { get; set; }
-
     public override IOperation GetOperation(IServiceProvider sp) => ActivatorUtilities.CreateInstance<MigrateRepoOperation>(sp, this);
-
-    public override Task<IServiceCollection> RegisterServices(IServiceCollection services)
-    {
-        services.AddSingleton(PcsApiFactory.GetAuthenticated(
-            PcsUri,
-            accessToken: null,
-            managedIdentityId: null,
-            disableInteractiveAuth: false));
-
-        TmpPath ??= Path.GetTempPath();
-
-        services.AddMultiVmrSupport(TmpPath);
-
-        return base.RegisterServices(services);
-    }
 }
